Correct use of wrong variable in mustache template
@climber-girl noticed a situation where the table of results in a
stakeholder's Trustworthy Email report contained checkmarks in the
"Has DMARC Policy of Reject" column, but the overview section
stated that 0% of hosts has such a policy.

Digging into this, I found that the mustache template that produces
the table of results was using the variable valid_dmarc_policy_reject,
where it should have been using the variable
valid_dmarc_policy_of_reject.  This commit fixes that.