Make changes to speed up report generation
The long pole in the tent is the retrieval of PTR records for the IPs
associated with DMARC aggregate reports.

First I added a least-recently-used cache for these lookups.  This
does not help much because the IPs don't tend to repeat, but it should
help a little.

Second I reduced the timeout for the PTR record lookup from the
default of 30 seconds to 2.5 seconds.  I noticed that many of the IPs
that appear in the DMARC aggregate reports consume the entire timeout,
while the lookups tend to return pretty quickly if they don't
timeout.  Thus this change considerably reduces the burden due to
unresponsive DNS hosts.