Fix incorrect logic
The logic that determines whether or not a DMARC report represents an
actual failure was incorrect.  It was counting the report as a failure
if EITHER DKIM OR SPF failed or failed alignment.  The correct
behavior is to count the report as a failure if BOTH DKIM AND SPF fail
or fail alignment.