Move query to avoid cursor timeout
We have been seeing the Trustymail report for HHS fail the past few
weeks.  It appears that the cursor is timing out and closing after ten
minutes of inactivity, as described in the docs:
https://api.mongodb.com/python/current/api/pymongo/collection.html#pymongo.collection.Collection.find

This is because the self.__db.sslyze_scan.find() query is taking more
than ten minutes to complete.  We can fix this by moving the cursor
creation code after the self.__db.sslyze_scan.find() query; that way,
the cursor has no chance to time out.