Move entire background LaTeX package config into \backgroundsetup
Previously the configuration was spread between \backgroundsetup and
the square braces of \usepackage[...]{background}.

Also move \backgroundsetup to just before we actually use
\BgThispage.  I think some of the LaTeX wizardry that comes after the
preamble but before the table of contents is messing up the way the
background LaTeX package works.

Whatever the reason, making the two changes described above the one
described above appears to fix a bug (described in #52) that has
popped up with our recent move to Debian Bullseye in #51 where the
opacity specified for the DHS logo on the table of contents page is
ignored.

When I tried only making the first change the DHS logo would show up
with full opacity on the title page and then appear correctly on the
table of contents, which is also undesirable.  It may be possible to
get the desired behavior with only the second change, but the first
change is good to make for consistency's sake.

The watermarks are known to me acting a little wonky.  I created
cisagov/trustymail_reporter#54 to document this.