Use str() casts to avoid None values
This past weekend we had a case where a DKIM result was None.  This
made the " ".join([y["result"] for y in dkim]) line fail and caused a
report not to be generated.  By using an explicit cast we causes None
values to be treated as "None" and therefore avoid this issue.